train sudormrf

+++ enhancement/sisdr_loss.py
@@ -112,9 +112,9 @@ class PermInvariantSISDR(nn.Module):
                                 permuted_pr_batch,
                                 t_batch,
                                 t_t_diag, eps=10e-8):
-        s_t = (self.dot(permuted_pr_batch, t_batch) /
-               (t_t_diag + eps) * t_batch)
-        e_t = permuted_pr_batch - s_t
+        s_t = ((self.dot(permuted_pr_batch, t_batch) /
+               (t_t_diag + eps) * t_batch)).float()
+        e_t = (permuted_pr_batch - s_t).float()
         sisnrs = 10 * torch.log10(self.dot(s_t, s_t) /
                                   (self.dot(e_t, e_t) + eps))
         return sisnrs
